+++ cms/fields/index.js
@@ -1,11 +1,16 @@
 export const ID = { label: 'ID', name: 'id', widget: 'uuid' };
 
-export const VariantField = (initial, options = []) => ({
-  label: 'Variant',
-  name: 'variant',
+export const SelectField = (
+  initial,
+  options = [],
+  label = 'Variant',
+  name = 'variant',
+) => ({
+  label,
+  name,
   widget: 'select',
-  default: initial,
-  options: options,
+  options,
+  ...(initial ? { default: initial } : null),
 });
 
 export const ImageField = (name = 'image', fieldName = 'photo') => ({
@@ -36,7 +41,7 @@ export const Button = {
       widget: 'string',
       required: false,
     },
-    VariantField('default', ['default', 'arrow', 'button', 'outlined']),
+    SelectField('default', ['default', 'arrow', 'button', 'outlined']),
   ],
 };
 
@@ -59,4 +64,38 @@ export const Content = {
   name: 'content',
   widget: 'markdown',
   required: false,
+};
+
+export const SettingsGroup = {
+  label: 'Settings',
+  name: 'settings',
+  widget: 'object',
+  collapsed: true,
+  fields: [
+    SelectField('default', ['default', 'dark', 'gray']),
+    SelectField(
+      'md',
+      ['none', 'sm', 'md', 'lg', 'xl'],
+      'Padding Top',
+      'padding_top',
+    ),
+    SelectField(
+      'md',
+      ['none', 'sm', 'md', 'lg', 'xl'],
+      'Padding Bottom',
+      'padding_bottom',
+    ),
+    SelectField(
+      'none',
+      ['none', 'sm', 'md', 'lg', 'xl'],
+      'Margin Top',
+      'margin_top',
+    ),
+    SelectField(
+      'none',
+      ['none', 'sm', 'md', 'lg', 'xl'],
+      'Margin Bottom',
+      'margin_bottom',
+    ),
+  ],
 };