+++ src/blocks/RecentArticles.js
@@ -1,16 +1,20 @@
-import React from 'react';
-import Recent from '../components/Article/Recent';
-import Container from '../components/UI/Container';
+import React from 'react'
+import Recent from '@/components/Article/Recent'
+import Container from '@/components/UI/Container'
+import Section from '@/components/UI/Section'
 
 export default function RecentArticles({ data, preview }) {
   return (
-    <section className='py-10 dark:bg-black dark:text-white lg:py-16'>
+    <Section
+      settings={data?.settings}
+      className="py-10 dark:bg-black dark:text-white lg:py-16"
+    >
       <Container>
-        <h2 className='mb-12 text-lg font-semibold uppercase tracking-wider md:text-xl'>
+        <h2 className="mb-12 text-lg font-semibold uppercase tracking-wider md:text-xl">
           {data?.title}
         </h2>
         {preview ? 'Articles will show up here' : <Recent />}
       </Container>
-    </section>
-  );
+    </Section>
+  )
 }