+++ src/components/Footer.js
@@ -1,15 +1,16 @@
-import React from 'react';
+import React from 'react'
 
 export default function Footer() {
   return (
-    <footer className='mt-auto py-5 bottom-0 w-full dark:bg-black bg-white dark:text-white text-black'>
-      <div className='container text-center mx-auto'>
+    <footer className="bottom-0 mt-auto w-full bg-white py-5 text-black dark:bg-black dark:text-white">
+      <div className="container mx-auto text-center">
         <a
-          href='https://cleancommit.io?utm_source=HenloGatsby&utm_medium=referral'
-          className='text-center text-sm font-regular text-gray-400 hover:text-gray-500 transition-colors'>
+          href="https://cleancommit.io?utm_source=HenloGatsby&utm_medium=referral"
+          className="font-regular text-center text-sm text-gray-400 transition-colors hover:text-gray-500"
+        >
           Created by cleancommit.io
         </a>
       </div>
     </footer>
-  );
+  )
 }