+++ tailwind.config.js
@@ -1,3 +1,4 @@
+/** @type {import('tailwindcss').Config} */
 module.exports = {
   darkMode: 'class',
   content: ['./src/**/*.{js,jsx,ts,tsx}'],
@@ -5,4 +6,4 @@ module.exports = {
     extend: {},
   },
   plugins: [require('@tailwindcss/typography'), require('@tailwindcss/forms')],
-};
+}